HomeWorkDev2-16

+++ src/main/java/org/example/homeworkdev214/controller/NoteController.java
@@ -1,19 +1,17 @@
 package org.example.homeworkdev214.controller;
 
 import org.example.homeworkdev214.model.Note;
+import org.example.homeworkdev214.model.dto.*;
 import org.example.homeworkdev214.service.NoteService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-@Controller
+@RestController
+@RequestMapping("/api/v1/notes")
 public class NoteController {
+
     private final NoteService noteService;
 
     @Autowired
@@ -21,29 +19,42 @@ public class NoteController {
         this.noteService = noteService;
     }
 
-    @GetMapping("/note/list")
-    public String listNotes(Model model) {
-        List<Note> notes = noteService.listAll();
-        model.addAttribute("notes", notes);
-        return "noteList";
+    @PostMapping
+    public CreateNoteResponse create(@RequestBody CreateNoteRequest request) {
+        Note note = new Note();
+        note.setTitle(request.getTitle());
+        note.setContent(request.getContent());
+        Note createdNote = noteService.add(note);
+        return createdNote != null ? CreateNoteResponse.success(createdNote.getId()) :
+                CreateNoteResponse.failed(CreateNoteResponse.Error.invalidTitle); // Можна додати детальну валідацію
     }
 
-    @PostMapping("/note/delete")
-    public String deleteNote(@RequestParam("id") long id) {
-        noteService.deleteById(id);
-        return "redirect:/note/list";
+    @GetMapping
+    public List<Note> getNotes() {
+        return noteService.listAll();
     }
 
-    @GetMapping("/note/edit")
-    public String editNoteForm(@RequestParam("id") long id, Model model) {
+    @PatchMapping("/{id}")
+    public UpdateNoteResponse update(@PathVariable long id, @RequestBody UpdateNoteRequest request) {
         Note note = noteService.getById(id);
-        model.addAttribute("note", note);
-        return "noteEdit";
+        if (note == null) {
+            return UpdateNoteResponse.failed(UpdateNoteResponse.Error.invalidNoteId);
+        }
+
+        note.setTitle(request.getTitle());
+        note.setContent(request.getContent());
+        noteService.update(note);
+
+        return UpdateNoteResponse.success(note);
     }
 
-    @PostMapping("/note/edit")
-    public String editNoteSubmit(@ModelAttribute Note note) {
-        noteService.update(note);
-        return "redirect:/note/list";
+    @DeleteMapping("/{id}")
+    public DeleteNoteResponse delete(@PathVariable long id) {
+        Note note = noteService.getById(id);
+        if (note == null) {
+            return DeleteNoteResponse.failed(DeleteNoteResponse.Error.invalidNoteId);
+        }
+        noteService.deleteById(id);
+        return DeleteNoteResponse.success();
     }
 }
